refactor: AST Linterをライブラリ化し、関心事ごとにモジュールを分割。マジックナンバーの定数化とビルドルールの警告修正、カバレッジゲートの調整を含む。
