+++ crates/katana-linter/Cargo.toml
@@ -5,13 +5,11 @@ edition.workspace = true
 license.workspace = true
 description = "KatanA AST Linter: workspace-wide static analysis for coding standards enforcement"
 
-[[test]]
-name = "ast_linter"
-path = "ast_linter.rs"
-
-[dev-dependencies]
+[dependencies]
 syn.workspace = true
 ignore.workspace = true
 serde_json.workspace = true
 proc-macro2 = { version = "1", features = ["span-locations"] }
+
+[dev-dependencies]
 tempfile.workspace = true
