+++ crates/katana-linter/src/rules/mod.rs
@@ -1,0 +1,5 @@
+pub mod changelog;
+pub mod i18n;
+pub mod locales;
+pub mod markdown;
+pub mod rust;
