+++ crates/katana-ui/Cargo.toml
@@ -26,6 +26,7 @@ dirs = "6.0.0"
 sha2 = "0.10"
 regex = "1.12.3"
 chrono = "0.4.44"
+tempfile = "3"
 
 [build-dependencies]
 cc = "1"
@@ -44,5 +45,4 @@ short_description = "A fast, lightweight Markdown workspace for macOS"
 
 [dev-dependencies]
 egui_kittest = { version = "0.33", features = ["snapshot", "wgpu", "eframe"] }
-tempfile = "3"
 time = "0.3"
